requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
PyJWT==2.8.0
Werkzeug==2.3.7
python-socketio==5.9.0
python-engineio==4.7.1
openai>=1.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
beautifulsoup4==4.12.2
